@@ -24,23 +24,89 @@
  * Contact the current copyright holder to discuss commerical license options.
  */
 
-using Microsoft.AspNetCore.Hosting;
+using Crypter.API.Models;
+using Crypter.API.Services;
+using Crypter.API.Startup;
+using Crypter.Core;
+using Crypter.Core.Interfaces;
+using Crypter.Core.Models;
+using Crypter.Core.Services.DataAccess;
+using Crypter.CryptoLib.Services;
+using Hangfire;
+using Hangfire.PostgreSql;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace CrypterAPI
-{
-   public class Program
-   {
-      public static void Main(string[] args)
-      {
-         CreateHostBuilder(args).Build().Run();
-      }
+var builder = WebApplication.CreateBuilder(args);
 
-      public static IHostBuilder CreateHostBuilder(string[] args) =>
-          Host.CreateDefaultBuilder(args)
-              .ConfigureWebHostDefaults(webBuilder =>
-              {
-                 webBuilder.UseStartup<Startup>();
-              });
-   }
+builder.Services.AddSingleton<ISimpleEncryptionService, SimpleEncryptionService>();
+builder.Services.AddSingleton<ISimpleHashService, SimpleHashService>();
+builder.Services.AddSingleton<ISimpleSignatureService, SimpleSignatureService>();
+builder.Services.AddSingleton<ITokenService, TokenService>();
+builder.Services.AddScoped<IEmailService, EmailService>();
+builder.Services.AddScoped<IApiValidationService, ApiValidationService>();
+
+builder.Services.AddDbContext<DataContext>();
+builder.Services.AddScoped<IUserService, UserService>();
+builder.Services.AddScoped<IUserProfileService, UserProfileService>();
+builder.Services.AddScoped<IUserPrivacySettingService, UserPrivacySettingService>();
+builder.Services.AddScoped<IUserPublicKeyPairService<UserX25519KeyPair>, UserX25519KeyPairService>();
+builder.Services.AddScoped<IUserPublicKeyPairService<UserEd25519KeyPair>, UserEd25519KeyPairService>();
+builder.Services.AddScoped<IUserSearchService, UserSearchService>();
+builder.Services.AddScoped<IUserEmailVerificationService, UserEmailVerificationService>();
+builder.Services.AddScoped<IUserNotificationSettingService, UserNotificationSettingService>();
+builder.Services.AddScoped<IUserTokenService, UserTokenService>();
+builder.Services.AddScoped<IBaseTransferService<IMessageTransferItem>, MessageTransferItemService>();
+builder.Services.AddScoped<IBaseTransferService<IFileTransferItem>, FileTransferItemService>();
+builder.Services.AddScoped<ISchemaService, SchemaService>();
+
+var configuration = builder.Configuration;
+var tokenSettings = configuration.GetSection("TokenSettings").Get<TokenSettings>();
+builder.Services.AddSingleton((serviceProvider) => tokenSettings);
+builder.Services.AddSingleton((serviceProvider) => configuration.GetSection("EmailSettings").Get<EmailSettings>());
+
+builder.Services.AddHangfire(config => config.UsePostgreSqlStorage(configuration.GetConnectionString("HangfireConnection")));
+builder.Services.AddHangfireServer(options => options.WorkerCount = configuration.GetValue<int>("HangfireSettings:Workers"));
+
+builder.Services.AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
+   .AddJwtBearerConfiguration(tokenSettings);
+
+builder.Services.AddCors();
+builder.Services.AddControllers();
+
+builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddSwaggerGen();
+
+var app = builder.Build();
+
+if (app.Environment.IsDevelopment())
+{
+   app.UseDeveloperExceptionPage();
+   app.UseCors(x =>
+      x.AllowAnyMethod()
+      .AllowAnyHeader()
+      .AllowAnyOrigin());
+
+   app.UseSwagger();
+   app.UseSwaggerUI();
+
+   using var serviceScope = app.Services.GetService<IServiceScopeFactory>()!.CreateScope();
+   var context = serviceScope.ServiceProvider.GetRequiredService<DataContext>();
+   context.Database.EnsureCreated();
 }
+
+app.UseHttpsRedirection();
+app.UseRouting();
+app.UseAuthentication();
+app.UseAuthorization();
+
+app.UseEndpoints(endpoints =>
+{
+   endpoints.MapControllers();
+   endpoints.MapHangfireDashboard();
+});
+
+app.Run();
