@@ -28,11 +28,11 @@ namespace Crypter.API.Models
 {
    public class EmailSettings
    {
-      public bool Enabled { get; set; }
-      public string From { get; set; }
-      public string Username { get; set; }
-      public string Password { get; set; }
-      public string Host { get; set; }
-      public int Port { get; set; }
+      public bool Enabled { get; set; } = false;
+      public string From { get; set; } = string.Empty;
+      public string Username { get; set; } = string.Empty;
+      public string Password { get; set; } = string.Empty;
+      public string Host { get; set; } = string.Empty;
+      public int Port { get; set; } = 0;
    }
 }
